Cleanup qc variable attribute renaming
