+++ tsdat/transform_v2/converters/nearest_neighbor.py
@@ -3,10 +3,10 @@ from typing import TYPE_CHECKING, Any, Optional
 import xarray as xr
 
 from ...io.base import DataConverter, RetrievedDataset
+from ...utils.replace_qc_attr import replace_qc_attr
 from ..nearest_neighbor.calculate_nearest_neighbor import nearest_neighbor
 from ..utils.is_metric_var import is_metric_var
 from ..utils.is_qc_var import is_qc_var
-from ..utils.replace_qc_attr import replace_qc_attr
 
 # Prevent any chance of runtime circular imports for typing-only imports
 if TYPE_CHECKING:  # pragma: no cover
